Updated for the latest version of roc
