+++ README.md
@@ -1,5 +1,7 @@
-# roc-plugin-style-sass
-[![Build Status](https://travis-ci.org/rocjs/roc-plugin-style-sass.svg?branch=master)](https://travis-ci.org/rocjs/roc-plugin-style-sass)
+# roc-plugin-style-sass [![Build Status](https://travis-ci.org/rocjs/roc-plugin-style-sass.svg?branch=master)](https://travis-ci.org/rocjs/roc-plugin-style-sass)
 
 __Plugin providing Sass support for Webpack__  
-- [roc-plugin-style-sass](/packages/roc-plugin-style-sass)
+- [roc-plugin-style-sass](/extensions/roc-plugin-style-sass)
+
+---
+To be used together with [Roc](https://github.com/rocjs/roc).
