+++ extensions/roc-plugin-style-sass/docs/Configuration.md
@@ -1,0 +1,5 @@
+# Config for `roc-plugin-style-sass`
+
+Configuration that can be defined in `roc.config.js`, other than settings, init and actions.
+
+__No config available.__
