+++ extensions/roc-plugin-style-sass/docs/Dependencies.md
@@ -1,0 +1,12 @@
+# Dependencies for `roc-plugin-style-sass`
+
+The dependencies that are available in the project.
+
+## Exported
+Nothing is exported.
+
+## Uses
+Nothing is listed as used.
+
+## Requires
+Nothing is required.
