+++ extensions/roc-plugin-style-sass/docs/Extensions.md
@@ -1,0 +1,7 @@
+# Extensions for `roc-plugin-style-sass`
+
+The extensions that are used in the project, indirect and direct.
+## Packages
+No packages.
+## Plugins
+_No plugins._
