+++ extensions/roc-plugin-style-sass/docs/Hooks.md
@@ -1,0 +1,19 @@
+# Hooks for `roc-plugin-style-sass`
+
+## Hooks
+* [roc](#roc)
+  * [update-settings](#update-settings)
+
+## roc
+
+### update-settings
+
+Expected to return new settings that should be merged with the existing ones.
+
+__Initial value:__ _Nothing_  
+__Expected return value:__ `{}`
+
+#### Arguments
+| Name        | Description                                                                  | Type       | Required | Can be empty |
+| ----------- | ---------------------------------------------------------------------------- | ---------- | -------- | ------------ |
+| getSettings | A function that returns the settings after the context has been initialized. | `Function` | No       | Yes          |
