+++ extensions/roc-plugin-style-sass/docs/Settings.md
@@ -1,0 +1,8 @@
+# Settings for `roc-plugin-style-sass`
+
+## Build
+
+### Sass
+| Name       | Description                                                                                        | Path                  | CLI option              | Default | Type      | Required | Can be empty | Extensions            |
+| ---------- | -------------------------------------------------------------------------------------------------- | --------------------- | ----------------------- | ------- | --------- | -------- | ------------ | --------------------- |
+| useBourbon | If Bourbon should be made available to include easily with &#96;@import &quot;bourbon&quot;;&#96;. | build.sass.useBourbon | --build-sass-useBourbon | `true`  | `Boolean` | Yes      | Yes          | roc-plugin-style-sass |
