+++ extensions/roc-plugin-style-sass/package.json
@@ -1,0 +1,45 @@
+{
+  "name": "roc-plugin-style-sass",
+  "version": "1.0.0-beta.0",
+  "description": "Plugin providing Sass support for Webpack in Roc",
+  "main": "lib/index.js",
+  "scripts": {
+    "lint": "eslint .",
+    "test": "npm run lint"
+  },
+  "files": [
+    "lib"
+  ],
+  "keywords": [
+    "roc",
+    "roc-plugin",
+    "webpack",
+    "style",
+    "sass"
+  ],
+  "repository": "https://github.com/rocjs/roc-plugin-style-sass/tree/master/extensions/roc-plugin-style-sass",
+  "author": "Verdens Gang AS",
+  "contributors": [{
+    "name": "Gustaf Dalemar"
+  }],
+  "license": "MIT",
+  "dependencies": {
+    "bourbon": "~4.2.6",
+    "node-sass": "~3.4.2",
+    "roc": "^1.0.0-rc",
+    "sass-loader": "~3.1.2",
+    "webpack": "~1.12.14"
+  },
+  "devDependencies": {
+    "babel-eslint": "~6.1.2",
+    "eslint": "~3.0.1",
+    "eslint-config-airbnb-base": "~4.0.0",
+    "eslint-plugin-babel": "~3.3.0",
+    "eslint-plugin-import": "~1.10.2"
+  },
+  "roc": {
+    "packages": [
+      "./lib/index.js"
+    ]
+  }
+}
