--- packages/roc-plugin-style-sass/.eslintrc
@@ -1,13 +1,0 @@
-{
-    "extends": "vgno",
-
-    "parser": "babel-eslint",
-
-    "env": {
-        "es6": true
-    },
-
-    "ecmaFeatures": {
-        "modules": true
-    }
-}
