--- packages/roc-plugin-style-sass/README.md
@@ -1,8 +1,0 @@
-# roc-plugin-style-sass
-Plugin providing Sass support for Webpack.
-
-## Documentation
-- [Actions](/packages/roc-plugin-style-sass/docs/Actions.md)
-- [Commands](/packages/roc-plugin-style-sass/docs/Commands.md)
-- [Hooks](/packages/roc-plugin-style-sass/docs/Hooks.md)
-- [Settings](/packages/roc-plugin-style-sass/docs/Settings.md)
