--- packages/roc-plugin-style-sass/docs/Hooks.md
@@ -1,1 +1,0 @@
-No hooks available.
