--- packages/roc-plugin-style-sass/docs/Settings.md
@@ -1,9 +1,0 @@
-# Settings for `roc-plugin-style-sass`
-
-## Build
-
-### Sass
-
-| Name       | Description                                                                                        | Path                  | CLI option              | Default | Type      | Required |
-| ---------- | -------------------------------------------------------------------------------------------------- | --------------------- | ----------------------- | ------- | --------- | -------- |
-| useBourbon | If Bourbon should be made available to include easily with &#96;@import &quot;bourbon&quot;;&#96;. | build.sass.useBourbon | --build-sass-useBourbon | `true`  | `Boolean` | No       |
