--- packages/roc-plugin-style-sass/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "roc-plugin-style-sass",
-  "version": "1.0.0-alpha.1",
-  "description": "Plugin providing Sass support for Webpack in Roc",
-  "main": "lib/index.js",
-  "scripts": {
-    "lint": "eslint .",
-    "test": "npm run lint"
-  },
-  "files": [
-    "lib"
-  ],
-  "keywords": [
-    "roc",
-    "roc-package",
-    "webpack",
-    "roc-dev"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/rocjs/roc-plugin-style-sass.git"
-  },
-  "author": "VG",
-  "license": "MIT",
-  "dependencies": {
-    "bourbon": "~4.2.6",
-    "node-sass": "~3.4.2",
-    "roc": "^1.0.0-rc",
-    "sass-loader": "~3.1.2",
-    "webpack": "~1.12.14"
-  },
-  "devDependencies": {
-    "babel-eslint": "~5.0.0",
-    "eslint": "~1.10.3",
-    "eslint-config-vgno": "~5.0.0"
-  }
-}
