--- packages/roc-plugin-style-sass/roc.config.js
@@ -1,6 +1,0 @@
-const path = require('path');
-
-// Makes it possible for use to generate documentation for this package.
-module.exports = {
-    packages: [path.join(__dirname, 'lib', 'index.js')]
-};
